requirements.txt
python-dateutil>=2.7
numpy~=1.23.1
scipy>=1.9.0
opencv-python==4.5.5.64
pyproj==3.3.1